+++ NEWS.md
@@ -1,4 +1,6 @@
 # aion 1.0.1.9000
+## Bugfixes & changes
+* Rename the `format` argument of `format()` to `prefix`.
 
 # aion 1.0.1
 ## Bugfixes & changes